Complete Rework.
Due to needing to get more knowledge and the time crunch I reworked it to a simpler code base. Removal of the MathClasses that were a naive translation of C# to C++ which was badly using pointers.

In this version the game is currently working on an infinite loop. To be fixed is waves and enemy positioning.